+++ sdk/core/Azure.Core/src/Shared/DiagnosticScopeFactory.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading;
@@ -17,6 +19,8 @@ namespace Azure.Core.Pipeline
         private readonly string? _resourceProviderNamespace;
         private readonly DiagnosticListener? _source;
         private readonly bool _suppressNestedClientActivities;
+
+        private static readonly ConcurrentDictionary<string, object?> ActivitySources = new();
 
         public DiagnosticScopeFactory(string clientNamespace, string? resourceProviderNamespace, bool isActivityEnabled, bool suppressNestedClientActivities)
         {
@@ -47,13 +51,43 @@ namespace Azure.Core.Pipeline
             {
                 return default;
             }
-            var scope = new DiagnosticScope(_source.Name, name, _source, kind, _suppressNestedClientActivities);
+
+            var scope = new DiagnosticScope(
+                                scopeName: name,
+                                source: _source,
+                                diagnosticSourceArgs: null,
+                                activitySource: GetActivitySource(_source.Name, name),
+                                kind: kind,
+                                suppressNestedClientActivities: _suppressNestedClientActivities);
 
             if (_resourceProviderNamespace != null)
             {
                 scope.AddAttribute("az.namespace", _resourceProviderNamespace);
             }
             return scope;
+        }
+
+        /// <summary>
+        /// This method combines client namespace and operation name into an ActivitySource name and creates the activity source.
+        /// For example:
+        ///     ns: Azure.Storage.Blobs
+        ///     name: BlobClient.DownloadTo
+        ///     result Azure.Storage.Blobs.BlobClient
+        /// </summary>
+        private static object? GetActivitySource(string ns, string name)
+        {
+            if (!ActivityExtensions.SupportsActivitySource())
+            {
+                return null;
+            }
+
+            string clientName = ns;
+            int indexOfDot = name.IndexOf(".", StringComparison.OrdinalIgnoreCase);
+            if (indexOfDot != -1)
+            {
+                clientName += "." + name.Substring(0, indexOfDot);
+            }
+            return ActivitySources.GetOrAdd(clientName, static n => ActivityExtensions.CreateActivitySource(n));
         }
     }
 }
